INT-5751: Fix moving blocks back into empty regions
This use case for core is only really supported for dragging
block specifically from the side-pre/post to the other
side-pre/post.  Dragging a block from any other custom
region does not work.

To get around this, just show all regions when editing
is enabled.  This will allow the user to drag blocks into
them from anywhere.

Also, to improve dragging blocks to top region, made the
drop zone have a min height so it is easier to hit.